refactor: extract benchmark run ledger cli commands
