+++ loopx/cli_commands/__init__.py
@@ -20,6 +20,10 @@ from .benchmark_review_lifecycle import (
     handle_benchmark_review_lifecycle_command,
     register_benchmark_review_lifecycle_commands,
 )
+from .benchmark_run_ledger import (
+    handle_benchmark_run_ledger_command,
+    register_benchmark_run_ledger_commands,
+)
 from .doctor import handle_doctor_command, register_doctor_command
 from .dreaming import handle_dreaming_command, register_dreaming_commands
 from .history import handle_history_command, register_history_command
@@ -67,6 +71,7 @@ __all__ = [
     "handle_agents_last_exam_command",
     "handle_benchmark_boundary_command",
     "handle_benchmark_review_lifecycle_command",
+    "handle_benchmark_run_ledger_command",
     "handle_check_command",
     "handle_codex_cli_bounded_visible_pilot_adapter_command",
     "handle_codex_cli_bootstrap_message_command",
@@ -100,6 +105,7 @@ __all__ = [
     "register_agents_last_exam_commands",
     "register_benchmark_boundary_commands",
     "register_benchmark_review_lifecycle_commands",
+    "register_benchmark_run_ledger_commands",
     "register_doctor_command",
     "register_dreaming_commands",
     "register_history_command",
